Resolve COMMON-3005 "Feat/"

See merge request web/ecom/ui!284

+++ src/components/Uploader/Uploader.js
@@ -55,7 +55,10 @@ export default function Uploader({
   }
 
   function removeImage() {
-    onRemove()
+    if (onRemove) {
+      onRemove()
+    }
+
     setIsLoaded(false)
     setImgSrc("")
   }
